src/requirements.txt
appdirs==1.4.3
attrs==16.3.0
Cython==0.25.2
Kivy==1.9.1
Kivy-Garden==0.1.4
packaging==16.8
pygame==1.9.3
pyparsing==2.2.0
pysendfile==2.0.1
requests==2.13.0
six==1.10.0
tqdm==4.11.2
